Change the type system, again!
FunT now takes two type variables `a/val` and `c/ctx`. The first
parameter of TyForall takes `a` as the sole argument, while the
second and the third parameters of TyForall takes `c` as the
argument.

The idea of `c` is to accomodate the case where the expression is
`pure`: if the expression is pure, then the second and the third
parameter should just return c as it is.

If we have a complex usage of type variables, we can nest them in
multiple `TyForall`. For the inner-most expression with complicated
manipulation of type variables, we use the `a`s to type it. Along the
path, we use the `c`s to type the 'context' - marking that the expression,
before being fully instantiated with necessary types, is pure.