Simplify the strategy to decrease fuel.
Fuel will always decrease on call, and will never increase even after we exit
the call. In a sense, fuel is global.

The reason for this is because all calls are "linearized" by continuations. We
can achieve "backtracking" of fuel if we implement the interpreter with a meta
language continuation (a function of type val -> imonad iresult). However, the
extraction of such interpreter is not as nice.