Finish optimizing the interpreter
I have found a way to invoke a captured continuation of `control` or `shallow handle` in
an elegant manner. The idea is simple: just promote `append/app` into a constructor of
`kont` which allows a O(1) append.

A cherry on top is that, this change allows us to write `interpret_kont_aux_app` and
`refine_kont_app`, in which `interpret_kont_aux_app` is O(n) where n is the size of the
continuation. Also, `interpreter_kont_aux_app` is not evaluated if we do not provide a
value, which means that `refine_kont_app` is O(1).

This brings huge benefit to the runtime: `reverse` example, on a list of 1000 elements,
now takes 300ms instead of 600ms. That's 2x speed up!