Remove faulty implement of `catch` and add `shift/reset` for `ces_monad`.
It seems like it is impossible to implement `catch` for `ces_monad` with the desired semantics and with
the signature `m a -> (e -> m a) -> m a`.

The reason why the old implementation is faulty is that, we catch error at the wrong time: only error
raised by `m` should be caught, but we catch error in `run m k`, wence includes error raised in the
continuation. To achieve the desired semantics, it is likely that we need to modify the signature of
catch to:
```
catch :: ces_monad a e s a -> (e -> ces_monad r e s a) -> ces_monad r e s a
```
And then pass a "delimited continuation" to `m` to "delimit" the scope of `catch`. But this signature
is not the standard catch signature that we want!

Additionally, we generalize the signature of `shift` for the standard `cont`, and then implement
`shift` and `reset` for ces_monad. The delimited continuation captured by `shift` is PURE with
respect to the continuation effect, so for `cont`, type of the delimited continuation is `a -> r`,
which is fine. But for `ces_monad`, tyoe of the delimited continuation is `a -> e -> (e + r) * s`,
which is extremely awkward to read (even if we introduce type alias). Generalize the type of the
delimited continuation to `forall r'. a -> cont r' r` and `forall r'. a -> ces_monad r' e s r` is
much more consistent.

A small note: we can use "rank 2 type" by putting the `forall` inside the argument of `shift`, and
that would be faithful to `shift/reset`. However, that would hurts extraction (apparently Rocq cannot
handle rank-2 type very well, so we have to put the `forall` quantifier for `r'` at the top level (of
`shift`) instead of putting it at the argument